Move Retry buttons
