+++ src/components/widgets/UploadStepper/UploadResultsContent.js
@@ -12,8 +12,8 @@ import ErrorOutlineIcon from '@material-ui/icons/ErrorOutline';
 import { jsx } from '@emotion/core';
 import {actionCreators} from "../../../redux/actionCreators";
 import colors from "../../../styles/colors";
-import RetryUploadButton from "../RetryButtons/RetryUploadButton";
-import RetryStartUploadOverButton from "../RetryButtons/RetryStartUploadOverButton";
+import RetryUploadButton from "../buttons/RetryButtons/RetryUploadButton";
+import RetryStartUploadOverButton from "../buttons/RetryButtons/RetryStartUploadOverButton";
 
 /**
  * This component watches for the upload state to change and then renders the appropriate
