+++ src/components/widgets/buttons/RetryButtons/RetryDownloadButton.js
@@ -2,9 +2,9 @@
 import { Fragment, useState } from "react";
 import { useSelector, useDispatch } from "react-redux";
 import { jsx } from "@emotion/core";
-import textStyles from "../../../styles/text";
-import buttonStyles from "../../../styles/buttons";
-import { actionCreators } from "../../../redux/actionCreators";
+import textStyles from "../../../../styles/text";
+import buttonStyles from "../../../../styles/buttons";
+import { actionCreators } from "../../../../redux/actionCreators";
 import Button from "@material-ui/core/Button/Button";
 
 /**
