+++ src/components/widgets/buttons/RetryButtons/RetryUploadButton.js
@@ -2,11 +2,11 @@
 import { Fragment, useState } from "react";
 import { useSelector, useDispatch } from "react-redux";
 import { jsx } from "@emotion/core";
-import textStyles from "../../../styles/text";
-import buttonStyles from "../../../styles/buttons";
-import { actionCreators } from "../../../redux/actionCreators";
+import textStyles from "../../../../styles/text";
+import buttonStyles from "../../../../styles/buttons";
+import { actionCreators } from "../../../../redux/actionCreators";
 import Button from "@material-ui/core/Button/Button";
-import LeftSpinner from "../LeftSpinner";
+import LeftSpinner from "../../LeftSpinner";
 
 export default function RetryUploadButton({selectedFile, selectedDuplicate, setStepThreeContent}) {
   const classes = buttonStyles.RetryUpload();
